Fix the headers and includes arguments in the spio.cuda.nvrtc.Program intializer.
Modify the generate() function to return the generated code instead of writing it to a file.

Remove the GenDirectory class.

Modify Kernel to pass the generated parameters header as compile args instead of writing it to a temporary file.

Modify kernel tests to use generated headers as strings instead of writing to a file.